# docs/swagger.yaml
definitions:
...
  coursesmanagament.GetCreditRequest:
    properties:
      semester:
        type: integer
      year:
        type: integer
    required:
    - semester
    - year
    type: object
...
  coursesmanagament.UserUnRegisterCourseRequest:
    properties:
      course_id:
        type: string
      course_semester:
        type: integer
      course_year:
        type: integer
    required:
    - course_id
    type: object
  departments.CreateDepartmentRequest:
    properties:
      department_code:
        maxLength: 10
        minLength: 2
        type: string
      department_name:
        maxLength: 100
        minLength: 3
        type: string
    required:
    - department_code
    - department_name
    type: object
  departments.UpdateDepartmentRequest:
    properties:
      department_code:
        type: string
      department_id:
        type: string
      department_name:
        type: string
    required:
    - department_code
    - department_id
    - department_name
    type: object
  global.SetTuitionTypeRequest:
    properties:
      cost:
        type: integer
      type:
        type: string
    required:
    - cost
    - type
    type: object
...
  models.Department:
    properties:
      createdAt:
        type: string
      departmentCode:
        type: string
      departmentName:
        type: string
      id:
        type: string
      updatedAt:
        type: string
    type: object
  models.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: boolean
    type: object
...
  tuition.PayTuitionRequest:
    properties:
      pay:
        type: integer
      semester:
        type: integer
      tuition_id:
        type: string
      year:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /department/create:
    post:
      consumes:
      - application/json
      description: Create department
      parameters:
      - description: CreateDepartmentRequest
        in: body
        name: CreateDepartmentRequest
        required: true
        schema:
          $ref: '#/definitions/departments.CreateDepartmentRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Department'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Create department
      tags:
      - Department
...
  /department/update:
    put:
      consumes:
      - application/json
      description: Update department
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: UpdateDepartmentRequest
        in: body
        name: UpdateDepartmentRequest
        required: true
        schema:
          $ref: '#/definitions/departments.UpdateDepartmentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update department
      tags:
      - Department
  /global/state:
    post:
      description: Set global state
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: State
        enum:
        - freeze
        - active
        - setup
        - done
        in: query
        name: state
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Set global state
      tags:
      - Global
  /global/tuition_type:
    get:
      description: Get tuition type
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get tuition type
      tags:
      - Global
    post:
      description: Set tuition type
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: SetTuitionTypeRequest
        in: body
        name: SetTuitionTypeRequest
        required: true
        schema:
          $ref: '#/definitions/global.SetTuitionTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Set tuition type
      tags:
      - Global
...
  /tuition/pay:
    post:
      consumes:
      - application/json
      description: Pay tuition
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: PayTuitionRequest
        in: body
        name: PayTuitionRequest
        required: true
        schema:
          $ref: '#/definitions/tuition.PayTuitionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Pay tuition
      tags:
      - Tuition
...
  /user/course/unregister:
    post:
      consumes:
      - application/json
      description: User unregister course
      parameters:
      - description: UserUnRegisterCourseRequest
        in: body
        name: UserUnRegisterCourseRequest
        required: true
        schema:
          $ref: '#/definitions/coursesmanagament.UserUnRegisterCourseRequest'
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: User unregister course
      tags:
      - User
  /user/credit:
    post:
      consumes:
      - application/json
      description: Get credit
      parameters:
      - default: Bearer <Add access token here>
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: GetCreditRequest
        in: body
        name: GetCreditRequest
        required: true
        schema:
          $ref: '#/definitions/coursesmanagament.GetCreditRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get credit
      tags:
      - User
...
swagger: "2.0"